[MATH] pow using 64bit

+++ math/math.h
@@ -10,11 +10,11 @@ extern "C" {
 #define INFINITY __builtin_inff()
 #endif
 
-static inline int64_t powi(int n, int e){
+static inline i64 powi(i64 n, i64 e){
     if (e == 1) return n;
     if (e == 2) return n << e;
-    int acc = n;
-    for (int i = 1; i < e; i++)
+    i64 acc = n;
+    for (i64 i = 1; i < e; i++)
         acc *= n;
     return acc;
 }
